requirements.txt
mysql-connector-python==9.3.0
openpyxl==3.1.5
pandas==2.3.1
python-dotenv==1.1.1
sqlalchemy==2.0.41
unidecode==1.4.0
chardet<5
python-dateutil<=2.9.0
